+++ taskpanel/src/components.d.ts
@@ -8,9 +8,13 @@ export {}
 declare module 'vue' {
   export interface GlobalComponents {
     AppFooter: typeof import('./components/AppFooter.vue')['default']
+    FilterSearch: typeof import('./components/FilterSearch.vue')['default']
     HelloWorld: typeof import('./components/HelloWorld.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    Settings: typeof import('./components/Settings.vue')['default']
+    SettingsButton: typeof import('./components/SettingsButton.vue')['default']
     TaskCard: typeof import('./components/TaskCard.vue')['default']
+    TaskSelector: typeof import('./components/TaskSelector.vue')['default']
   }
 }
