+++ taskpanel/src/components/TaskSelector.vue
@@ -1,65 +1,11 @@
 <template>
-  <v-container fluid class="taskpanel" class="fill-height d-flex flex-column">
+  <v-container fluid class="taskpanel fill-height d-flex flex-column">
     <v-row
       v-if="search"
       id="filters"
       class="align-self-start d-flex w-100 flex-grow-0"
     >
-      <v-col cols="9" sm="4" md="7">
-        <v-text-field
-          id="search"
-          v-model="search"
-          clearable
-          label="Search"
-          hint="Search for tasks by name or description"
-        ></v-text-field>
-      </v-col>
-      <v-col cols="3" class="mt-2 d-flex d-sm-none">
-        <v-btn variant="tonal">
-          <v-icon>mdi-cog</v-icon>
-          <v-menu activator="parent">
-            <v-list>
-              <v-list-item
-                v-for="setting in userSettings"
-                :title="setting.title"
-                @click="setting.action"
-              ></v-list-item>
-            </v-list>
-          </v-menu>
-        </v-btn>
-      </v-col>
-      <v-col cols="6" sm="3" md="2">
-        <v-combobox
-          clearable
-          chips
-          multiple
-          label="Tags"
-          v-model="selectedTags"
-          :items="taskTags"
-        ></v-combobox>
-      </v-col>
-      <v-col cols="6" sm="3" md="2" class="d-flex align-self-start">
-        <v-switch
-          label="Show All"
-          inset
-          color="secondary"
-          v-model="showall"
-        ></v-switch>
-      </v-col>
-      <v-col cols="1" class="mt-2 d-none d-sm-flex">
-        <v-btn variant="tonal">
-          <v-icon>mdi-cog</v-icon>
-          <v-menu activator="parent">
-            <v-list>
-              <v-list-item
-                v-for="setting in userSettings"
-                :title="setting.title"
-                @click="setting.action"
-              ></v-list-item>
-            </v-list>
-          </v-menu>
-        </v-btn>
-      </v-col>
+      <FilterSearch :items="tasks" @filter="filterTasks"></FilterSearch>
     </v-row>
     <v-row
       id="main-content"
@@ -68,7 +14,7 @@
       class="d-flex flex-row w-100"
     >
       <TaskCard
-        v-for="task in tasks"
+        v-for="task in tasklist"
         :task="task"
         :working="workingdata[task.ID]"
         :selected="selected == task.ID"
@@ -90,10 +36,6 @@
 
 <script lang="ts" setup>
 const props = defineProps({
-  search: {
-    type: Boolean,
-    required: false,
-  },
   tasks: {
     type: Array,
     required: true,
@@ -106,9 +48,15 @@ const props = defineProps({
     type: Number,
     required: false,
   },
+  search: {
+    type: Boolean,
+    required: false,
+  },
 });
 
-const selectTask = (taskID: number) => {
+const tasklist = ref(props.tasks);
+
+const selectTask = (taskID: Number) => {
   if (props.selected == taskID) {
     return;
   }
@@ -118,4 +66,12 @@ const selectTask = (taskID: number) => {
     emit("select", 0);
   }
 };
+
+const filterTasks = (tasks) => {
+  tasklist.value = tasks;
+};
+
+const emit = defineEmits<{
+  (e: "select", taskID: Number): void;
+}>();
 </script>
